Changed python bindings from ctypes to cffi.
The reason is that I want to bundle the aduana library with the python bindings.
The best I can do with ctypes is put it inside the egg file, which can give me
a file stream to the library, but to open it with ctypes I need a path...
All solutions with ctypes, except installing the library inside the system, look like
a hack.